Added ROOT_SIMPLE output. This saves events in a ROOT tree format where list mode variables are simple numbers, with branch names identifying board and channel. Waveforms, if present, are stored in a vector. Only the first (earliest) hit in a group of coincidences is taken.
ALso added the option to specify '--help', '-help', or '-h' as a command line arg.
This produces the usage message.

Note that the ROOT_SIMPLE part is not complete. Right now, the code cannot
determine the full set of board, channel combinations. For the moment, a placeholder
of 0,4 and 0,5 is used.